[UPDATE] refactoration in paths so they make sense
[RENAME] files and diretories
