+++ src/components/Dashboards/DashboardDetail.vue
